requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
cloudinary==1.25.0
cryptography==3.4.8
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.8.0
django-decouple==2.1
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djongo==1.3.6
gunicorn==20.1.0
idna==2.10
mongoengine==0.23.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pymongo==3.12.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.4
whitenoise==5.2.0
